+++ src/io/slice_io.cc
@@ -256,7 +256,7 @@ namespace io {
 				throw __EXCEPT__(FILE_WRITE_ERROR);
 			}
 
-			if (fwrite(&frame.slice_cnt, sizeof(size_t), 1, fp) != 1) {
+			if (fwrite(&frame.slice_cnt, sizeof(uint32_t), 1, fp) != 1) {
 				throw __EXCEPT__(FILE_WRITE_ERROR);
 			}
 
@@ -268,7 +268,7 @@ namespace io {
 				throw __EXCEPT__(FILE_WRITE_ERROR);
 			}
 
-			if (fwrite(frame.size.data(), sizeof(size_t), frame.slice_cnt, fp) != frame.slice_cnt) {
+			if (fwrite(frame.size.data(), sizeof(uint32_t), frame.slice_cnt, fp) != frame.slice_cnt) {
 				throw __EXCEPT__(FILE_WRITE_ERROR);
 			}
 
@@ -280,7 +280,7 @@ namespace io {
 				throw __EXCEPT__(FILE_WRITE_ERROR);
 			}
 
-			if (fwrite(frame.geometry_size.data(), sizeof(size_t), frame.slice_cnt, fp) != frame.slice_cnt) {
+			if (fwrite(frame.geometry_size.data(), sizeof(uint32_t), frame.slice_cnt, fp) != frame.slice_cnt) {
 				throw __EXCEPT__(FILE_WRITE_ERROR);
 			}
 			for (auto& i : frame.geometry) {
@@ -292,7 +292,7 @@ namespace io {
 				}
 			}
 
-			if (fwrite(frame.color_size.data(), sizeof(size_t), frame.slice_cnt, fp) != frame.slice_cnt) {
+			if (fwrite(frame.color_size.data(), sizeof(uint32_t), frame.slice_cnt, fp) != frame.slice_cnt) {
 				throw __EXCEPT__(FILE_WRITE_ERROR);
 			}
